+++ CHANGELOG.md
@@ -1821,3 +1821,4 @@
 - [2026-06-06 20:30] feat(state): implement responsive state view component
 - [2026-06-06 20:51] test(ui): add unit tests for ui business logic
 - [2026-06-06 20:56] style(state): harmonize light and dark theme palette for state
+- [2026-06-06 20:59] test(auth): implement widget integration tests for auth